+++ core/metal/demosaic_bayer_linear.metal
@@ -6,7 +6,7 @@
 #include "shader_common.h"
 
 kernel void demosaic_bayer_linear(
-    const device uint16_t* raw_data [[buffer(0)]],
+    const device float* raw_data [[buffer(0)]],
     device float* rgb_data [[buffer(1)]],
     constant DemosaicBayerParams& params [[buffer(2)]],
     uint2 gid [[thread_position_in_grid]]
@@ -23,7 +23,7 @@ kernel void demosaic_bayer_linear(
     const uint rgb_idx = raw_idx * 3;
     const uint pixel_color = fcol_bayer(row, col, filters);
     
-    const float raw_val = float(raw_data[raw_idx]) / params.maximum_value;
+    const float raw_val = raw_data[raw_idx] / params.maximum_value;
     
     rgb_data[rgb_idx + 0] = 0.0f;
     rgb_data[rgb_idx + 1] = 0.0f;
@@ -44,7 +44,7 @@ kernel void demosaic_bayer_linear(
                 uint ny = row + dy_diag[i];
                 uint nx = col + dx_diag[i];
                 if (fcol_bayer(ny, nx, filters) == c) {
-                    sum += float(raw_data[ny * width + nx]) / params.maximum_value;
+                    sum += raw_data[ny * width + nx] / params.maximum_value;
                     count++;
                 }
             }
@@ -57,7 +57,7 @@ kernel void demosaic_bayer_linear(
                         uint ny = row + dy;
                         uint nx = col + dx;
                         if (fcol_bayer(ny, nx, filters) == c) {
-                            sum += float(raw_data[ny * width + nx]) / params.maximum_value;
+                            sum += raw_data[ny * width + nx] / params.maximum_value;
                             count++;
                         }
                     }
